refactor: Reorganize adapters into separate directory structure (Phase 1)
This commit reorganizes the engine's adapter code to align with
Architecture v2 design. Changes include:

- Move src/io/node/* to src/adapters/node/
- Move src/io/memory/* to src/adapters/memory/
- Create adapter index files for cleaner imports
- Create placeholder directories for future adapters (daemon, browser)
- Create presets directory for future preset functions
- Update all import paths throughout codebase

All tests pass (vitest + Python integration tests).
This is Phase 1 of the Architecture v2 implementation plan.
