+++ packages/engine/src/adapters/memory/index.ts
@@ -1,0 +1,2 @@
+export { InMemoryFileSystem } from './memory-filesystem'
+export { MemorySocket, MemorySocketFactory } from './memory-socket'
