+++ packages/engine/src/adapters/node/index.ts
@@ -1,0 +1,4 @@
+export { NodeFileSystem } from './node-filesystem'
+export { NodeSocketFactory, NodeTcpSocket, NodeUdpSocket } from './node-socket'
+export { NodeStorageHandle } from './node-storage-handle'
+export { ScopedNodeFileSystem } from './scoped-node-filesystem'
