+++ packages/engine/src/adapters/node/scoped-node-filesystem.ts
