+++ packages/engine/test/core/torrent.spec.ts
@@ -3,7 +3,7 @@ import { describe, it, expect, beforeEach } from 'vitest'
 import { Torrent } from '../../src/core/torrent'
 import { PieceManager } from '../../src/core/piece-manager'
 import { TorrentContentStorage } from '../../src/core/torrent-content-storage'
-import { InMemoryFileSystem } from '../../src/io/memory/memory-filesystem'
+import { InMemoryFileSystem } from '../../src/adapters/memory'
 import { BitField } from '../../src/utils/bitfield'
 import { PeerConnection } from '../../src/core/peer-connection'
 import { ITcpSocket } from '../../src/interfaces/socket'
