+++ packages/engine/test/integration/memory-swarm.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach } from 'vitest'
 import { BtEngine } from '../../src/core/bt-engine'
-import { MemorySocketFactory } from '../../src/io/memory/memory-socket'
-import { InMemoryFileSystem } from '../../src/io/memory/memory-filesystem'
+import { MemorySocketFactory } from '../../src/adapters/memory'
+import { InMemoryFileSystem } from '../../src/adapters/memory'
 import { TorrentCreator } from '../../src/core/torrent-creator'
 import { PeerConnection } from '../../src/core/peer-connection'
 import { ISocketFactory, ITcpSocket, IUdpSocket } from '../../src/interfaces/socket'
