+++ packages/engine/test/integration/node-download.spec.ts
@@ -7,9 +7,9 @@ import { Torrent } from '../../src/core/torrent'
 import { PieceManager } from '../../src/core/piece-manager'
 import { TorrentContentStorage } from '../../src/core/torrent-content-storage'
 import { BitField } from '../../src/utils/bitfield'
-import { NodeSocketFactory } from '../../src/io/node/node-socket'
-// import { NodeFileSystem } from '../../src/io/node/node-filesystem'
-import { NodeStorageHandle } from '../../src/io/node/node-storage-handle'
+import { NodeSocketFactory } from '../../src/adapters/node'
+// import { NodeFileSystem } from '../../src/adapters/node'
+import { NodeStorageHandle } from '../../src/adapters/node'
 import { PeerConnection } from '../../src/core/peer-connection'
 import { PeerWireProtocol, MessageType } from '../../src/protocol/wire-protocol'
 import { MockEngine } from '../utils/mock-engine'
