+++ packages/engine/test/tracker/udp-tracker-integration.spec.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeAll, afterAll } from 'vitest'
 import { BtEngine } from '../../src/core/bt-engine'
 import { TorrentCreator } from '../../src/core/torrent-creator'
-import { ScopedNodeFileSystem } from '../../src/io/node/scoped-node-filesystem'
-import { NodeSocketFactory } from '../../src/io/node/node-socket'
-import { NodeStorageHandle } from '../../src/io/node/node-storage-handle'
+import { ScopedNodeFileSystem } from '../../src/adapters/node'
+import { NodeSocketFactory } from '../../src/adapters/node'
+import { NodeStorageHandle } from '../../src/adapters/node'
 import { SimpleTracker } from '../helpers/simple-tracker'
 import * as crypto from 'crypto'
 import path from 'path'
